Improve exports and usage
