+++ zui/src/index.ts
@@ -1,11 +1,12 @@
-export type {
-  BaseType,
-  GlobalComponentDefinitions,
-  UIComponentDefinitions,
-  SchemaResolversMap,
-  AsBaseType,
-} from './ui/types'
-export { type ZuiFormProps, ZuiForm, schemaToUISchema } from './ui'
+import { UIComponentDefinitions } from './ui/types'
+export type { BaseType, UIComponentDefinitions, SchemaResolversMap, ZuiComponentMap, AsBaseType } from './ui/types'
+export {
+  type ZuiFormProps,
+  ZuiForm,
+  schemaToUISchema,
+  transformZuiComponentsToRenderers,
+  defaultUISchemaResolvers,
+} from './ui'
 export type { Zui, ZuiType, Infer, ZuiExtension, ZuiRawShape, ZuiTypeAny } from './zui'
 export type {
   JsonSchema7Type as JsonSchema7,
@@ -16,3 +17,11 @@ export { zui } from './zui'
 export { getZuiSchemas } from './zui-schemas'
 export { jsonSchemaToZui } from './json-schema-to-zui'
 export { ZodError as ZuiError } from 'zod'
+
+export interface ComponentDefinitions {}
+
+export type GlobalComponentDefinitions = ComponentDefinitions extends {
+  components: infer TComponentMap extends UIComponentDefinitions
+}
+  ? TComponentMap
+  : any
